refactor: Move CheckCommand() to its own file
